Update for BluetoothConnector 2.0
As the workflow didn't work for me, I guess due to another Version of BluetoothConnector it couldn't retrieve the device list, I changed it a bit. BluetoothConnector 2.0 gives an error when no argument is provided. So I had to catch the stderr and to change the regex to avoid listing 'Error:' and 'MAC' in deviceMacs. For me it works perfect now.